+++ tests/Integration/Core/Reflection/ReflectionTraitTest.php
@@ -3,9 +3,9 @@
 namespace Phpactor\WorseReflection\Tests\Integration\Core\Reflection;
 
 use Phpactor\WorseReflection\Tests\Integration\IntegrationTestCase;
-use Phpactor\WorseReflection\Bridge\TolerantParser\Reflection\ReflectionTrait;
 use Phpactor\WorseReflection\Core\ClassName;
 use Phpactor\WorseReflection\Core\Type;
+use Phpactor\WorseReflection\Core\Reflection\ReflectionTrait;
 
 class ReflectionTraitTest extends IntegrationTestCase
 {
