+++ tests/Unit/Core/DocblockTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Phpactor\WorseReflection\Tests\Unit\Core\Reflection\Formatted;
+namespace Phpactor\WorseReflection\Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
 use Phpactor\WorseReflection\Core\Docblock;
 
-class DocblockTest extends TestCase
+class Core extends TestCase
 {
     /**
      * @testdox It returns true for "none" if no docblock is present.
