+++ tests/Unit/Core/Inference/LocalAssignmentsTest.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Phpactor\WorseReflection\Tests\Unit\Core\Reflection\Inference;
+namespace Phpactor\WorseReflection\Tests\Unit\Core\Inference;
 
 use Phpactor\WorseReflection\Core\Inference\Assignments;
 use Phpactor\WorseReflection\Core\Inference\LocalAssignments;
+use Phpactor\WorseReflection\Tests\Unit\Core\Inference\AssignmentstTestCase;
 
 class LocalAssignmentsTest extends AssignmentstTestCase
 {
